Implement M1a (Netcode + Sim) — encrypted reliable-UDP, full §8.5 handshake, ECS sim, 89/89 tests green
Adds the complete M1a milestone: platform-independent netcode and simulation
verified on Linux (89 tests), Windows-specific drivers (CNG, Winsock,
ERServer, ERHeadless) wired up and ready to build on MSVC.

Networking layer (NeuronCore/net/):
- Protocol.h: wire contract (magic, version, channels, message types, header)
- SequenceMath.h: 16-bit wraparound comparison/diff for reliable sequencing
- ReplayWindow.h: 1024-bit sliding window (R13 anti-replay)
- Reliability.h: RTT/RTO (RFC 6298), Karn's algorithm, ack+bitfield
- Fragmentation.h: Fragmentize() + Reassembler (~1200B safe payload)
- PacketCodec.h: little-endian header encode/decode; message framing
- ICrypto.h / ISocket.h: abstract crypto and socket contracts
- SecureChannel.h: AES-256-GCM AEAD framing + replay window
- HandshakeMessages.h: wire structs for all 9 handshake message types
- Handshake.h: HandshakeServerStateless + HandshakeServerConnection +
  HandshakeClient with idempotency guards (m_cookieHandled, m_haveKeys)
  to survive packet duplication/reorder
- FakeCrypto.h: deterministic fake (sorted-pair ECDH, real AEAD semantics)
  for platform-independent test coverage
- CngCrypto.h/.cpp: Windows CNG production impl (ECDH P-256, ECDSA P-256,
  AES-256-GCM, HKDF, PBKDF2-SHA512, HMAC-SHA256 stateless cookie)
- WinsockSocket.h/.cpp: non-blocking UDP; dual-stack IPv6/IPv4 fallback
- LoopbackNetwork.h: in-memory network with loss/dup/reorder impairments
- Connection.h: ClientConnection + ServerConnection state machines
- ServerHost.h: cookie phase, per-connection map, snapshot broadcast

Simulation layer (NeuronCore/sim/):
- Components.h: Transform, Velocity, NetId, BaseTag, ShipTag, Health; enums
- Movement.h: IntegrateMovement (int64 fold-back), ClampSpeed, MovementSystem
- Snapshot.h: SnapshotEntity + versioned Encode/DecodeSnapshot
- Command.h: MoveCommand + Encode/DecodeMoveCommand
- ServerWorld.h: SpawnBase (200m before boundary, 40 m/s east), Step,
  BuildSnapshot; SpawnPlayer assigns PlayerNetId for client tracking
- SimComponents.cpp: single NEURON_DEFINE_COMPONENT site for ERServer/ERHeadless

ECS fix (NeuronCore/ecs/Ecs.h):
- Changed `constexpr uint64_t required` to `const uint64_t` in both
  ForEach overloads — ComponentId::value is runtime static, not constexpr

ERServer / ERHeadless:
- ERServer.cpp: 30 Hz loop (CngCrypto + WinsockSocket + ServerHost +
  ServerWorld + FixedStepAccumulator); publishes er_server_pub.bin
- ERHeadless.cpp: N-bot host; loads pinned key, drives bots through §8.5,
  watches sector crossings (M1a acceptance scenario)
- Both vcxprojs updated to compile CngCrypto.cpp, WinsockSocket.cpp,
  SimComponents.cpp; ERServer gains odbc32 (M5 stub), ERHeadless bcrypt

IOCP skeleton (ERServer/netio/):
- IocpUdpListener.h/.cpp: M4 scaling placeholder; per-connection affinity
  comments; not wired into M1a main loop

Test suites added (NeuronTools/testrunner/):
- Tests_Net.h: 22 cases — SequenceMath, ReplayWindow, Reliability,
  Fragmentation, PacketCodec, Protocol
- Tests_Crypto.h: 10 cases — FakeCrypto ECDH, AEAD, cookie, PBKDF2
- Tests_Loopback.h: 5 cases — BasicSendRecv, Loss, Dup, Reorder, Seed
- Tests_Sim.h: Movement + Snapshot suites (defines all 6 component IDs)
- Tests_Handshake.h: 7 cases — FullSequence, SecureChannel, ReplayRejected,
  TamperedCiphertext, MitmSignature, VersionMismatch, BadCookie
- Tests_Integration.h: 4 cases — ThreeBots{Clean,Dup,Reorder,LossReorderDup}
  (M1a acceptance criterion: 3 bots cross sector boundary under impairments)

All 89 tests verified green on Linux (g++ -std=c++23).

+++ NeuronTools/testrunner/Tests_Math.h
@@ -13,45 +13,45 @@ TEST_SUITE(Math)
         XMVECTOR v = Neuron::Math::Set(3.0f, 4.0f, 0.0f);
         XMVECTOR n = Neuron::Math::Normalize(v);
         const float len = Neuron::Math::Length(n);
-        CHECK(std::fabsf(len - 1.0f) < 1e-6f);
+        CHECK(std::fabs(len - 1.0f) < 1e-6f);
     });
 
     TEST_CASE(Dot) {
         XMVECTOR a = Neuron::Math::Set(1.0f, 0.0f, 0.0f);
         XMVECTOR b = Neuron::Math::Set(0.0f, 1.0f, 0.0f);
         const float d = Neuron::Math::Dotf(a, b);
-        CHECK(std::fabsf(d) < 1e-6f); // perpendicular → 0
+        CHECK(std::fabs(d) < 1e-6f); // perpendicular → 0
     });
 
     TEST_CASE(Cross) {
         XMVECTOR x = Neuron::Math::Set(1.0f, 0.0f, 0.0f);
         XMVECTOR y = Neuron::Math::Set(0.0f, 1.0f, 0.0f);
         XMVECTOR z = Neuron::Math::Cross(x, y);
-        CHECK(std::fabsf(Neuron::Math::GetX(z))        < 1e-6f);
-        CHECK(std::fabsf(Neuron::Math::GetY(z))        < 1e-6f);
-        CHECK(std::fabsf(Neuron::Math::GetZ(z) - 1.0f) < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::GetX(z))        < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::GetY(z))        < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::GetZ(z) - 1.0f) < 1e-6f);
     });
 
     TEST_CASE(LengthSquare) {
         XMVECTOR v = Neuron::Math::Set(1.0f, 2.0f, 2.0f);
-        CHECK(std::fabsf(Neuron::Math::LengthSquare(v) - 9.0f) < 1e-6f);
-        CHECK(std::fabsf(Neuron::Math::Length(v)       - 3.0f) < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::LengthSquare(v) - 9.0f) < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::Length(v)       - 3.0f) < 1e-6f);
     });
 
     TEST_CASE(SetLength) {
         XMVECTOR v = Neuron::Math::Set(1.0f, 0.0f, 0.0f);
         XMVECTOR s = Neuron::Math::SetLength(v, 5.0f);
-        CHECK(std::fabsf(Neuron::Math::Length(s) - 5.0f) < 1e-5f);
+        CHECK(std::fabs(Neuron::Math::Length(s) - 5.0f) < 1e-5f);
     });
 
     TEST_CASE(SmoothStep01) {
-        CHECK(std::fabsf(Neuron::Math::SmoothStep01(0.0f))        < 1e-6f);
-        CHECK(std::fabsf(Neuron::Math::SmoothStep01(1.0f) - 1.0f) < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::SmoothStep01(0.0f))        < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::SmoothStep01(1.0f) - 1.0f) < 1e-6f);
         // Mid-point: smooth-step(0.5) = 0.5
-        CHECK(std::fabsf(Neuron::Math::SmoothStep01(0.5f) - 0.5f) < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::SmoothStep01(0.5f) - 0.5f) < 1e-6f);
         // Clamped outside [0,1]
-        CHECK(std::fabsf(Neuron::Math::SmoothStep01(-1.0f))        < 1e-6f);
-        CHECK(std::fabsf(Neuron::Math::SmoothStep01(2.0f) - 1.0f)  < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::SmoothStep01(-1.0f))        < 1e-6f);
+        CHECK(std::fabs(Neuron::Math::SmoothStep01(2.0f) - 1.0f)  < 1e-6f);
     });
 
     TEST_CASE(AlignUp) {
